Major refactor for interupt driven operation. Also uses updated TVout library from my repository.

+++ jafar_sketch/rx5808.h
@@ -34,8 +34,8 @@ class RX5808
 {
   public:
     RX5808(uint16_t RSSIpin, uint16_t CSpin);
-    uint16_t getVal(uint16_t band, uint16_t channel, uint8_t norm);
-    uint16_t getVal(uint16_t pos, uint8_t norm);
+    uint16_t getVal(uint8_t band, uint8_t channel, uint8_t norm);
+    uint16_t getVal(uint8_t pos, uint8_t norm);
     uint16_t getMaxPosBand(uint8_t band);
     uint16_t getMaxValBand(uint8_t band, uint8_t norm);
     uint16_t getMinPosBand(uint8_t band);
@@ -46,7 +46,7 @@ class RX5808
     void calibration();
     void setFreq(uint32_t freq);
     void abortScan();
-    uint16_t getfrom_top8(uint8_t index);
+    uint8_t getfrom_top8(uint8_t index);
     void compute_top8(void);
     uint16_t getRssi(uint16_t channel);
     uint16_t getCurrentRSSI();
@@ -59,7 +59,7 @@ class RX5808
     uint16_t _csPin;
     uint8_t _stop_scan;
     uint16_t scanVec[CHANNEL_MAX];
-    uint16_t scanVecTop8[8];
+    uint8_t scanVecTop8[8];
 
     void serialEnable(const uint8_t);
     void serialSendBit(const uint8_t);
